requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alive-progress==3.1.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
croniter==1.3.8
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
dnspython==2.3.0
elosports==0.1.1
email-validator==1.3.1
etils==1.1.0
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.16.3
filelock==3.10.6
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.6
jaxlib==0.4.6
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.11.0
keras-rl==0.4.2
kiwisolver==1.4.4
libclang==15.0.6.1
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.8.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.7
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
poke-env==0.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz==2023.2
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
readchar==4.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.3
rl==3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-macos==2.11.0
tensorflowjs==4.2.0
termcolor==2.2.0
terminado==0.17.1
tf-nightly-2.0-preview==2.0.1
tinycss2==1.2.1
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
